fix(algorithms,classification,widgets): resolve audit batch #429-#438 plus two master regressions (#439, #440)
- RoiStatisticsWidget: point-in-polygon masking + per-band NoData filtering
  instead of raw bounding-rectangle stats (#429)
- RsPostProcess::sieve: isolated components keep their original class when
  no bordering neighbors exist instead of being reset to class 0 (#430)
- TerrainAnalysis slope/aspect/hillshade: 1-sided border/NoData gradients
  divided by cellSize, not 2*cellSize (50% underestimate fixed) (#431)
- ImageFusion: size_t loop indices + 64-bit counters for large scenes (#432)
- RsRoiLabeler: reproject training geometries vector-CRS -> raster-CRS
  before rasterization, traditional GIS axis order, RAII handle (#433)
- BandMath: min/max/pow/atan2 propagate NaN symmetrically; processFile
  masks NoData via exact float compare (large sentinels like -3.4e38 now
  match); std::min/std::max expression aliases (#434)
- RadiometricCalibration: loadLandsatMtl/loadSentinel2Mtd auto-discover the
  band set from metadata when bandNames is empty, per documented contract;
  non-finite coefficient guards in kernels + processFile (#435)
- TransitionMatrix::marginals negative classCount guard; QaMask::sclMask
  null classes guard (#436)
- RsClassifierCvBackend fit/predict/save null m_clf guards (#437)
- RX anomaly: per-band hasNoData flags; no fabricated -9999 sentinel for
  rasters without NoData metadata (#438)
- stratifiedSplit: single-group classes fall back to pixel-level split so
  testSplit still yields held-out metrics (operator contract restored) (#439)
- toRadiance: reject identity defaults unless loaded (#301 semantics) while
  accepting explicit non-identity coefficients; resolves contradictory
  pinned tests (master suite was red) (#440)

Fixes #429
Fixes #430
Fixes #431
Fixes #432
Fixes #433
Fixes #434
Fixes #435
Fixes #436
Fixes #437
Fixes #438
Fixes #439
Fixes #440

+++ src/processing/algorithms/image_fusion.h
@@ -89,6 +89,6 @@ class ImageFusion
 
   private:
     /// Histogram-match src to ref (match mean and stddev).
-    static void histogramMatch( float *data, int n,
-                                const float *ref, int refN, float nodata );
+    static void histogramMatch( float *data, size_t n,
+                                const float *ref, size_t refN, float nodata );
 };
